+++ apps/api/filters/acquisitions.py
@@ -1,4 +1,5 @@
 import django_filters
+from django.db.models import Q
 
 from apps.core.models import Acquisition
 
@@ -20,6 +21,6 @@ class AcquisitionFilter(django_filters.FilterSet):
             return qs.none()
 
         if not self.request.user.is_superuser:
-            qs = qs.filter(user_acquisition__user=self.request.user)
+            qs = qs.filter(Q(entry__catalog__users=self.request.user) | Q(entry__catalog__is_public=True))
 
         return qs